refactor: 💡 change pacakge structure after feedback
- move config from top level 'bean' to new 'config' package
- move logger from top level 'bean' to new 'log' package
- move sentry funcs from top level 'bean' to 'trace' package
- add 'WithEcho' at the end of existing sentry capture func names
- create new funcs to log and capture with sentry
