+++ async/async.go
@@ -32,10 +32,11 @@ import (
 
 	"github.com/getsentry/sentry-go"
 	"github.com/labstack/echo/v4"
-	"github.com/retail-ai-inc/bean/v2"
+	"github.com/retail-ai-inc/bean/v2/config"
 	"github.com/retail-ai-inc/bean/v2/internal/gopool"
 	"github.com/retail-ai-inc/bean/v2/internal/regex"
-	bsentry "github.com/retail-ai-inc/bean/v2/internal/sentry"
+	"github.com/retail-ai-inc/bean/v2/log"
+	"github.com/retail-ai-inc/bean/v2/trace"
 )
 
 type (
@@ -64,7 +65,7 @@ func Execute(fn func(), poolName ...string) {
 				}
 			}
 		} else {
-			bean.Logger().Warnf("async func will execute without goroutine pool, the pool name is %q\n", poolName[0])
+			log.Logger().Warnf("async func will execute without goroutine pool, the pool name is %q\n", poolName[0])
 		}
 	}
 
@@ -75,7 +76,7 @@ func Execute(fn func(), poolName ...string) {
 // and provides all error stack aiming to facilitate fail causes discovery.
 func ExecuteWithContext(fn Task, c echo.Context, poolName ...string) {
 	functionName := "unknown function"
-	if bean.BeanConfig.Sentry.On && bean.BeanConfig.Sentry.TracesSampleRate > 0.0 {
+	if config.Bean.Sentry.On && config.Bean.Sentry.TracesSampleRate > 0.0 {
 		if pc, file, line, ok := runtime.Caller(1); ok {
 			functionName = fmt.Sprintf("%s:%d\n\t\r %s\n", path.Base(file), line, runtime.FuncForPC(pc).Name())
 		}
@@ -91,7 +92,7 @@ func ExecuteWithContext(fn Task, c echo.Context, poolName ...string) {
 		ctx := ec.Request().Context()
 		// IMPORTANT - Set the sentry hub key into the context so that `SentryCaptureException` and `SentryCaptureMessage`
 		// can pull the right hub and send the exception message to sentry.
-		if bean.BeanConfig.Sentry.On {
+		if config.Bean.Sentry.On {
 			hub := sentry.GetHubFromContext(ctx)
 			if hub == nil {
 				hub = sentry.CurrentHub().Clone()
@@ -100,7 +101,7 @@ func ExecuteWithContext(fn Task, c echo.Context, poolName ...string) {
 			hub.Scope().SetRequest(ec.Request())
 			ctx = sentry.SetHubOnContext(ctx, hub)
 
-			if bean.BeanConfig.Sentry.TracesSampleRate > 0.0 {
+			if config.Bean.Sentry.TracesSampleRate > 0.0 {
 				urlPath := ec.Request().URL.Path
 
 				span := sentry.StartSpan(ctx, "async",
@@ -131,7 +132,7 @@ func ExecuteWithContext(fn Task, c echo.Context, poolName ...string) {
 
 func ExecuteWithTimeout(ctx context.Context, duration time.Duration, fn TimeoutTask, poolName ...string) {
 	functionName := "unknown function"
-	if bean.BeanConfig.Sentry.On && bean.BeanConfig.Sentry.TracesSampleRate > 0.0 {
+	if config.Bean.Sentry.On && config.Bean.Sentry.TracesSampleRate > 0.0 {
 		if pc, file, line, ok := runtime.Caller(1); ok {
 			functionName = fmt.Sprintf("%s:%d\n\t\r %s\n", path.Base(file), line, runtime.FuncForPC(pc).Name())
 		}
@@ -160,7 +161,7 @@ func ExecuteWithTimeout(ctx context.Context, duration time.Duration, fn TimeoutT
 		}
 
 		// can pull the right hub and send the exception message to sentry.
-		if bean.BeanConfig.Sentry.On && bean.BeanConfig.Sentry.TracesSampleRate > 0.0 {
+		if config.Bean.Sentry.On && config.Bean.Sentry.TracesSampleRate > 0.0 {
 			var transactionName string
 			if parentSpan != nil {
 				transactionName = parentSpan.Name
@@ -183,14 +184,14 @@ func ExecuteWithTimeout(ctx context.Context, duration time.Duration, fn TimeoutT
 }
 
 func CaptureException(c context.Context, err error) {
-	bsentry.CaptureException(c, err, bean.Logger(), bean.BeanConfig.Sentry.On, "async context is missing hub information")
+	trace.SentryCaptureException(c, err)
 }
 
 // Recover the panic and send the exception to sentry.
 func recoverPanic(c context.Context) {
 	if err := recover(); err != nil {
 		// Create a new Hub by cloning the existing one.
-		if bean.BeanConfig.Sentry.On {
+		if config.Bean.Sentry.On {
 			var localHub *sentry.Hub
 
 			if c != nil {
@@ -208,6 +209,6 @@ func recoverPanic(c context.Context) {
 			localHub.Recover(err)
 		}
 
-		bean.Logger().Error(err)
+		log.Logger().Error(err)
 	}
 }
